라이크기능 추가 password 싹다 pw로 바꿈 모델에 데이터베이스 관계 추가

+++ repositories/likes.repository.js
@@ -1,4 +1,5 @@
 const { like } = require("../models");
+const { post } = require("../models");
 
 class LikeRepository {
     checkPost = async (postId) => {
@@ -11,16 +12,40 @@ class LikeRepository {
             return true;
         }
     };
+    postLikeUpDown= async(postId,likeNum) => {
+        try{const test =await post.findOne({
+            where: {postId}
+        })
+        
+        
+        await post.update(
+        {
+            like:test.like + likeNum
 
-    checkExistLikeAndDelete = async (postId, userId) => {
-        try {
-            await like.destroy({
-                where: { postId, userId },
-            });
-            return true;
-        } catch (err) {
-            return false;
+        },{
+            where : {postId}
         }
+        )}catch(err){
+            next(err)
+        }
+    }
+    checkExistLikeAndDelete = async (postId, userId) => {
+        
+        
+        const checkExistLikeAndDeleteData=await like.findOne({
+                where: {postId:Number(postId), userId },
+            });
+            
+            if(!checkExistLikeAndDeleteData === true){
+                return false;
+            }else {
+                
+                this.postLikeUpDown(postId,-1) 
+                await like.destroy({
+                where: { postId:Number(postId), userId },
+            })
+            return true;}
+        
     };
 
     createPostLike = async (postId, userId) => {
@@ -28,6 +53,7 @@ class LikeRepository {
             postId,
             userId,
         });
+        this.postLikeUpDown(postId,+1)
         return;
     };
 }
